dtype uint8

+++ BehaviorScreen/process.py
@@ -23,7 +23,7 @@ from BehaviorScreen.core import (
 def get_background_image(
         behavior_data: BehaviorData
     ) -> np.ndarray:
-    return np.asarray(behavior_data.metadata['background']['image_ROI']) 
+    return np.asarray(behavior_data.metadata['background']['image_ROI'], dtype=np.uint8) 
 
 def get_background_image_2(
         behavior_data: BehaviorData,
